# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Medclara Backend API
  description: |
    # Medclara Clinical Documentation Platform API
    
    Complete API specification for the Medclara backend, including:
    - **Authentication** - User login and session management
    - **Patients** - Patient data management
    - **Templates** - Clinical note templates
    - **Sessions** - Scribe session management
    - **Recordings** - Audio recording upload and processing
    - **Chunked Upload** - Large file upload with resumable chunks
    - **Server-Sent Events (SSE)** - Real-time processing status updates
    - **Notes** - Clinical note creation and management
    - **Workspace** - Scribe workspace API
    
    ## Real-Time Processing with SSE
    
    The Medclara API uses Server-Sent Events (SSE) for real-time recording processing updates.
    This provides 99% fewer network requests and 20x faster response times compared to polling.
    
    ### Complete Recording Workflow
    1. **POST** `/api/recordings/chunked/init` - Initialize upload session
    2. **POST** `/api/recordings/chunked/upload` - Upload audio chunks (repeat)
    3. **POST** `/api/recordings/chunked/finalize` - Start processing
    4. **GET** `/api/recordings/{recordingId}/stream` - Stream processing events (SSE)
    5. Process events: `processing` → `completed` or `failed`
    
    ### Benefits of SSE Over Polling
    - **99% fewer requests** - One persistent connection vs. polling every 500ms
...
    
  version: '1.0.0'
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code for programmatic handling
          examples:
            - unauthorized
            - not_found
            - invalid_request
            - server_error
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
...
    ChunkUploadResponse:
      type: object
      required:
        - chunkIndex
        - sessionId
        - chunksReceived
        - totalChunks
        - progress
      properties:
        chunkIndex:
          type: integer
          description: Index of uploaded chunk
        sessionId:
          type: string
          format: uuid
        chunksReceived:
          type: integer
        totalChunks:
          type: integer
        progress:
          type: number
          minimum: 0
          maximum: 100
        message:
          type: string
...
paths:
...
  /api/recordings/chunked/upload:
    post:
      tags:
        - Chunked Upload
      summary: Upload audio chunk
      description: |
        Upload a single chunk of audio data.
        
        Call this endpoint multiple times to upload all chunks.
        Each chunk should be sequentially numbered.
      operationId: uploadChunk
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - sessionId
                - chunkIndex
                - totalChunks
                - chunk
              properties:
                sessionId:
                  type: string
                  format: uuid
                chunkIndex:
                  type: integer
                  description: Chunk index (0-based)
                totalChunks:
                  type: integer
                  description: Total number of chunks
                isLastChunk:
                  type: boolean
                  description: Whether this is the last chunk
                chunk:
                  type: string
                  format: binary
                  description: Audio chunk data
      responses:
        '200':
          description: Chunk uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChunkUploadResponse'
        '400':
          description: Invalid chunk
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Chunk order invalid or out of sequence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
security:
  - BearerAuth: []